refactoring : #7 List<ItemInfo> -> ItemInfos - 일급 컬렉션으로 랩핑
