+++ src/main/java/me/hero/myrefrigerator/common/ItemInfos.java
@@ -1,0 +1,13 @@
+package me.hero.myrefrigerator.common;
+
+import me.hero.myrefrigerator.domain.dto.ItemInfo;
+
+import java.util.List;
+
+public class ItemInfos {
+    private final List<ItemInfo> itemInfos;
+
+    public ItemInfos(List<ItemInfo> itemInfos) {
+        this.itemInfos = itemInfos;
+    }
+}
