+++ src/main/java/me/hero/myrefrigerator/common/Items.java
@@ -1,11 +1,11 @@
-package me.hero.myrefrigerator.common.vo;
+package me.hero.myrefrigerator.common;
 
 import me.hero.myrefrigerator.domain.Item;
 
 import java.util.List;
 
 public class Items {
-    private List<Item> items;
+    private final List<Item> items;
 
     public Items(List<Item> items) {
         this.items = items;
